separate the non-react stuff into a controller
